// packages/shared/src/api/model/orderResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Drones API
 * OpenAPI spec version: v1
 */
import type { OrderItemResponse } from './orderItemResponse';

export interface OrderResponse {
  id?: string;
  userId?: string;
  restaurantId?: string;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  deliveryAddress?: string | null;
  deliveryLatitude?: number;
  deliveryLongitude?: number;
  totalAmount?: number;
  createdAt?: string;
  /** @nullable */
  updatedAt?: string | null;
  /** @nullable */
  items?: OrderItemResponse[] | null;
}